fix: deisng feedback and added tests
