+++ apps/journeys-admin/src/components/Editor/ActionDetails/ActionInformation/index.ts
@@ -1,0 +1,1 @@
+export { ActionInformation } from './ActionInformation'
