+++ apps/journeys-admin/src/components/Editor/ActionsTable/ActionsList/ActionsList.spec.tsx
@@ -6,8 +6,14 @@ import {
   ActiveFab,
   ActiveJourneyEditContent
 } from '@core/journeys/ui/EditorProvider'
+import useMediaQuery from '@mui/material/useMediaQuery'
 import { Actions } from '../ActionsTable'
 import { ActionsList } from './ActionsList'
+
+jest.mock('@mui/material/useMediaQuery', () => ({
+  __esModule: true,
+  default: jest.fn()
+}))
 
 jest.mock('@core/journeys/ui/EditorProvider', () => {
   const originalModule = jest.requireActual('@core/journeys/ui/EditorProvider')
@@ -52,20 +58,73 @@ describe('ActionsList', () => {
     })
   })
 
-  it('should render a list of actions', () => {
-    const { getAllByText } = render(
-      <ActionsList actions={actions} goalLabel={() => 'Visit a website'} />
+  describe('mdUp', () => {
+    beforeEach(() =>
+      (useMediaQuery as jest.Mock).mockImplementation(() => true)
     )
-    expect(getAllByText('https://www.google.com/')[0]).toBeInTheDocument()
-    expect(getAllByText('Visit a website')[0]).toBeInTheDocument()
-    expect(getAllByText(2)[0]).toBeInTheDocument()
+
+    it('should render the information drawer on the right', () => {
+      const { getByTestId, getByRole } = render(
+        <ActionsList actions={actions} goalLabel={() => 'Visit a website'} />
+      )
+      fireEvent.click(getByRole('button', { name: 'Learn More' }))
+      expect(getByTestId('ActionInformation').parentElement).toHaveClass(
+        'MuiDrawer-paperAnchorRight'
+      )
+    })
+
+    it('should close information drawer on close icon click', () => {
+      const { getByTestId, getByText, getByRole } = render(
+        <ActionsList actions={actions} goalLabel={() => 'Visit a website'} />
+      )
+      fireEvent.click(getByRole('button', { name: 'Learn More' }))
+      expect(getByText('Information')).toBeInTheDocument()
+      fireEvent.click(getByTestId('CloseIcon'))
+      expect(getByTestId('ActionInformation').parentElement).not.toHaveClass(
+        'MuiDrawer-parentAnchorRight'
+      )
+    })
+
+    it('should render a list of actions', () => {
+      const { getAllByText } = render(
+        <ActionsList actions={actions} goalLabel={() => 'Visit a website'} />
+      )
+      expect(getAllByText('https://www.google.com/')[0]).toBeInTheDocument()
+      expect(getAllByText('Visit a website')[0]).toBeInTheDocument()
+      expect(getAllByText(2)[0]).toBeInTheDocument()
+    })
+
+    it('should open the drawer or dispatch on click', () => {
+      const { getAllByTestId } = render(
+        <ActionsList actions={actions} goalLabel={() => 'Visit a website'} />
+      )
+      fireEvent.click(getAllByTestId('EditRoundedIcon')[0])
+      expect(dispatch).toHaveBeenCalled()
+    })
   })
 
-  it('should dispatch on click', () => {
-    const { getAllByTestId } = render(
-      <ActionsList actions={actions} goalLabel={() => 'Visit a website'} />
+  describe('mdDown', () => {
+    beforeEach(() =>
+      (useMediaQuery as jest.Mock).mockImplementation(() => false)
     )
-    fireEvent.click(getAllByTestId('EditRoundedIcon')[0])
-    expect(dispatch).toHaveBeenCalled()
+
+    it('should render the information drawer from the bottom', () => {
+      const { getByTestId, getByText, getByRole } = render(
+        <ActionsList actions={actions} goalLabel={() => 'Visit a website'} />
+      )
+      fireEvent.click(getByRole('button', { name: 'Learn More' }))
+      expect(getByText('Information')).toBeInTheDocument()
+      expect(getByTestId('ActionInformation').parentElement).toHaveClass(
+        'MuiPaper-root MuiPaper-elevation MuiPaper-elevation0 MuiDrawer-paper MuiDrawer-paperAnchorBottom css-1mdfdy2-MuiPaper-root-MuiDrawer-paper'
+      )
+    })
+
+    it('should render the actions list in mobile view', () => {
+      const { getAllByText } = render(
+        <ActionsList actions={actions} goalLabel={() => 'Visit a website'} />
+      )
+      expect(getAllByText('Target and Action')[0]).toBeInTheDocument()
+      expect(getAllByText('Appears on')).not.toHaveLength(2)
+    })
   })
 })
