+++ apps/journeys-admin/src/components/Editor/ActionsTable/ActionsList/ActionsListView/index.ts
@@ -1,0 +1,1 @@
+export { ActionsListView } from './ActionsListView'
