+++ apps/journeys-admin/src/components/Editor/EditToolbar/EditToolbar.tsx
@@ -2,10 +2,6 @@ import { ReactElement } from 'react'
 import IconButton from '@mui/material/IconButton'
 import VisibilityIcon from '@mui/icons-material/Visibility'
 import { useJourney } from '@core/journeys/ui/JourneyProvider'
-import {
-  ActiveJourneyEditContent,
-  useEditor
-} from '@core/journeys/ui/EditorProvider'
 import { JourneyStatus } from '../../../../__generated__/globalTypes'
 import { DuplicateBlock } from '../../DuplicateBlock'
 import { DeleteBlock } from './DeleteBlock'
@@ -13,9 +9,6 @@ import { Menu } from './Menu'
 
 export function EditToolbar(): ReactElement {
   const { journey } = useJourney()
-  const {
-    state: { journeyEditContentComponent }
-  } = useEditor()
 
   return (
     <>
@@ -23,11 +16,7 @@ export function EditToolbar(): ReactElement {
         aria-label="Preview"
         href={`/api/preview?slug=${journey?.slug ?? ''}`}
         target="_blank"
-        disabled={
-          journey == null ||
-          journey?.status === JourneyStatus.draft ||
-          journeyEditContentComponent === ActiveJourneyEditContent.Action
-        }
+        disabled={journey == null || journey?.status === JourneyStatus.draft}
       >
         <VisibilityIcon />
       </IconButton>
